+++ lookup_extensions/lookups/__init__.py
@@ -1,3 +1,3 @@
-from .exists import *  # noqa F401,F403
+from .complements import *  # noqa F401,F403
 from .exregex import *  # noqa F401,F403
 from .negate import *  # noqa F401,F403